pyproject.toml
[tool.poetry]
name = "sisneo-to-redcap"
version = "0.1.0"
description = "Exporta dados do Sisneo para o RedCap do HCPA"
authors = ["luanjpb <luanjuniopb@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
requests = "^2.31.0"
streamlit = "^1.28.0"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
requests-html = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
